add CMakeLists and fix bugs
